+++ BlockFactory/Shaders/Block/Vertex.glsl
@@ -2,7 +2,7 @@
 layout (location = 0) in vec3 aPos; // the position variable has attribute position 0
 layout (location = 1) in vec4 color;
 layout (location = 2) in vec2 uv;
-layout (location = 3) in int sprite;
+layout (location = 3) in float sprite;
 
 uniform mat4 model;
 uniform mat4 view;
@@ -22,5 +22,5 @@ void main()
     gl_Position = projection * view * model * vec4(aPos, 1.0); // see how we directly give a vec3 to vec4's constructor
     vertexColor = color;
     vertexUv = uv;
-    vertexSprite = sprite;
+    vertexSprite = int(sprite);
 }
